refactor: extract quote logic into hook
